do not use product flavor cuz it's very slow when testing
