--- DACMASDK-Sample/src/adpod/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="ad_tag_url"><![CDATA[http://webdemo.dac.co.jp/sdfactory/stsn/top.php?filepath=stsn/adpods/basic.xml]]></string>
-</resources>
