+++ DACMASDK-Sample/src/main/AndroidManifest.xml
@@ -21,6 +21,26 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity
+            android:name=".adpod.AdpodActivity"
+            android:label="@string/app_name" />
+
+        <activity
+            android:name=".vertical.VerticalActivity"
+            android:label="@string/app_name" />
+
+        <activity
+            android:name=".content.ContentActivity"
+            android:label="@string/app_name" />
+
+        <activity
+            android:name=".no_content.NoContentActivity"
+            android:label="@string/app_name" />
+
+        <activity
+            android:name=".vmap.VmapActivity"
+            android:label="@string/app_name" />
     </application>
 
 </manifest>
