+++ DACMASDK-Sample/src/main/java/jp/co/dac/sdk/ma/sample/AdCustomBinding.java
@@ -2,14 +2,13 @@ package jp.co.dac.sdk.ma.sample;
 
 import android.app.Activity;
 import android.content.Context;
-import android.databinding.BindingAdapter;
 import android.graphics.Point;
 import android.support.annotation.IntRange;
 import android.view.View;
 import android.view.ViewGroup;
 
 public class AdCustomBinding {
-    @BindingAdapter({"bind:verticalPercentage"})
+
     public static void setVerticalPercentage(View view, @IntRange(from = 0, to = 100) int verticalPercentage) {
         Context context = view.getContext();
         if (context instanceof Activity) {
