+++ DACMASDK-Sample/src/main/java/jp/co/dac/sdk/ma/sample/MainActivity.java
@@ -1,9 +1,13 @@
 package jp.co.dac.sdk.ma.sample;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
 import android.support.v7.app.AppCompatActivity;
+
+import jp.co.dac.sdk.ma.sample.adpod.AdpodActivity;
+import jp.co.dac.sdk.ma.sample.content.ContentActivity;
+import jp.co.dac.sdk.ma.sample.no_content.NoContentActivity;
+import jp.co.dac.sdk.ma.sample.vertical.VerticalActivity;
+import jp.co.dac.sdk.ma.sample.vmap.VmapActivity;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -12,13 +16,10 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(bundle);
         setContentView(R.layout.activity_main);
 
-        populateAdFragment(getString(R.string.ad_tag_url));
-    }
-
-    void populateAdFragment(String adTagUrl) {
-        Fragment fragment = AdFragment.newInstance(adTagUrl);
-        getSupportFragmentManager().beginTransaction()
-                .replace(R.id.container, fragment)
-                .commit();
+        findViewById(R.id.adpod).setOnClickListener(view -> startActivity(AdpodActivity.getCallingIntent(this)));
+        findViewById(R.id.vertical).setOnClickListener(view -> startActivity(VerticalActivity.getCallingIntent(this)));
+        findViewById(R.id.content).setOnClickListener(view -> startActivity(ContentActivity.getCallingIntent(this)));
+        findViewById(R.id.no_content).setOnClickListener(view -> startActivity(NoContentActivity.getCallingIntent(this)));
+        findViewById(R.id.vmap).setOnClickListener(view -> startActivity(VmapActivity.getCallingIntent(this)));
     }
 }
