+++ DACMASDK-Sample/src/main/res/layout/dac_ad_vertical_layout.xml
