+++ DACMASDK-Sample/src/main/res/layout/fragment_ad_vertical.xml
