+++ DACMASDK-Sample/src/main/res/values/strings.xml
@@ -1,4 +1,8 @@
 <resources>
     <string name="app_name">DAC MA SDK</string>
+
     <string name="ad_tag_url"><![CDATA[http://webdemo.dac.co.jp/sdfactory/stsn/top.php?filepath=stsn/unit/fruitsbear.xml]]></string>
+    <string name="ad_adpod_tag_url"><![CDATA[http://webdemo.dac.co.jp/sdfactory/stsn/top.php?filepath=stsn/adpods/basic.xml]]></string>
+    <string name="ad_vertical_tag_url"><![CDATA[http://xp1.zedo.com/jsc/xp2/fns.vast?n=2696&c=84/58&d=21&s=2&v=vast2&pu=__page-url__&ru=__referrer__&pw=__player-width__&ph=__player-height__&z=__random-number__]]></string>
+    <string name="ad_vmap_tag_url"><![CDATA[http://webdemo.dac.co.jp/sdfactory/stsn/top.php?filepath=stsn/vmap/pre-standalones2_mid15-1_post-1.xml]]></string>
 </resources>
