--- DACMASDK-Sample/src/vertical/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="ad_tag_url"><![CDATA[http://xp1.zedo.com/jsc/xp2/fns.vast?n=2696&c=84/58&d=21&s=2&v=vast2&pu=__page-url__&ru=__referrer__&pw=__player-width__&ph=__player-height__&z=__random-number__]]></string>
-</resources>
