+++ dependencies.gradle
@@ -2,7 +2,7 @@ ext {
     minSdkVersion = 15
     targetSdkVersion = 23
     compileSdkVersion = 23
-    buildToolsVersion = '23.0.3'
+    buildToolsVersion = '24.0.0 rc4'
     androidPlugin = 'com.android.tools.build:gradle:2.1.0'
 
     playService = 'com.google.android.gms:play-services-ads:8.4.+'
